Move nvt preferences components into an own module
